Atualizada regra de validaçao de bd - Criado mixin para nao misturar a logica da validacao com o resto do codigo
